Move new encoding tests to test_encoding.py
